+++ Modules/resources/Model/ReVisa.php
@@ -26,7 +26,7 @@ class ReVisa extends Model {
 		`id_resource_category` int(11) NOT NULL,
 		`id_instructor` int(11) NOT NULL,
 		`instructor_status` int(11) NOT NULL,
-        `is_active` int(0) NOT NULL,
+        `is_active` int(0) NOT NULL DEFAULT 1,
 		PRIMARY KEY (`id`)
 		);";
 
@@ -77,14 +77,14 @@ class ReVisa extends Model {
      * @return PDOStatement
      */
     public function createDefaultVisa($id_space) {
-        $sql = "insert into re_visas(id_resource_category, id_instructor, instructor_status, id_space)"
-                . " values(?,?,?, ?)";
+        $sql = "insert into re_visas(id_resource_category, id_instructor, instructor_status, id_space, is_active)"
+                . " values(?,?,?, ?, 1)";
         $this->runRequest($sql, array(0, 1, 1, $id_space));
     }
     
     public function importVisa($id_space, $id, $id_cat, $id_instructor, $instructor_status){
-        $sql = "insert into re_visas(id, id_resource_category, id_instructor, instructor_status, id_space)"
-                . " values(?,?,?,?,?)";
+        $sql = "insert into re_visas(id, id_resource_category, id_instructor, instructor_status, id_space, is_active)"
+                . " values(?,?,?,?,?,1)";
         $this->runRequest($sql, array($id, $id_cat, $id_instructor, $instructor_status, $id_space));
     }
 
@@ -133,8 +133,8 @@ class ReVisa extends Model {
      */
     public function addVisa($id_space, $id_resource_category, $id_instructor, $instructor_status) {
 
-        $sql = "insert into re_visas(id_resource_category, id_instructor, instructor_status, id_space)"
-                . " values(?,?,?,?)";
+        $sql = "insert into re_visas(id_resource_category, id_instructor, instructor_status, id_space, is_active)"
+                . " values(?,?,?,?,1)";
         $this->runRequest($sql, array($id_resource_category, $id_instructor, $instructor_status, $id_space));
         return $this->getDatabase()->lastInsertId();
     }
@@ -209,7 +209,7 @@ class ReVisa extends Model {
         if ($req->rowCount() == 1) {
             $visaInfo = $req->fetch();  // get the first line of the result
 
-            return $this->getVisaDesc($visaInfo, $lang);
+            return $this->getVisaDesc($id_space, $visaInfo, $lang);
         } else{
             return "";
         }