+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [booking] clicking on blank days in month view gives access to day area view
 * [projects] fix closed projects listing change of year
+* [resources] fix re_visa is_active default value
 
 ## 2.2.2